+++ R/gdalraster.R
@@ -359,17 +359,20 @@
 #' The dataset can be re-opened on the existing \code{filename} with 
 #' \code{$open(read_only=TRUE)} or \code{$open(read_only=FALSE)}.
 #'
-#' @seealso
-#' Package overview in [`help("gdalraster-package")`][gdalraster-package]
-#'
-#' [create()], [createCopy()], [rasterFromRaster()], [rasterToVRT()], [warp()]
-#'
-#' As described above, the `$read()` method will perform automatic resampling 
-#' if the specified output size (`out_xsize * out_ysize`) is different than 
+#' @note
+#' The `$read()` method will perform automatic resampling if the 
+#' specified output size (`out_xsize * out_ysize`) is different than 
 #' the size of the region being read (`xsize * ysize`). In that case, the 
 #' GDAL_RASTERIO_RESAMPLING configuration option could also be defined to 
 #' override the default resampling to one of BILINEAR, CUBIC, CUBICSPLINE, 
 #' LANCZOS, AVERAGE or MODE (see [set_config_option()]).
+#'
+#' @seealso
+#' Package overview in [`help("gdalraster-package")`][gdalraster-package]
+#'
+#' [create()], [createCopy()], [rasterFromRaster()], [rasterToVRT()]
+#'
+#' [fillNodata()], [warp()]
 #'
 #' @examples
 #' lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")