+++ src/gdal_exp.cpp
@@ -7,10 +7,11 @@
 #include <cmath>
 
 #include "gdal.h"
-#include "gdal_utils.h"
 #include "cpl_conv.h"
 #include "cpl_port.h"
 #include "cpl_string.h"
+#include "gdal_alg.h"
+#include "gdal_utils.h"
 
 #include <errno.h>
 
@@ -238,6 +239,8 @@ bool createCopy(std::string format, std::string dst_filename,
 		Rcpp::stop("Driver does not support create copy.");
 		
 	GDALDatasetH hSrcDS = GDALOpenShared(src_filename.c_str(), GA_ReadOnly);
+	if(hSrcDS == NULL)
+		Rcpp::stop("Open source raster failed.");
 	
 	std::vector<char *> opt_list = {NULL};
 	if (options.isNotNull()) {
@@ -385,6 +388,86 @@ Rcpp::IntegerMatrix get_pixel_line(const Rcpp::NumericMatrix xy,
 											inv_gt[5] * geo_y));
 	}
 	return pixel_line;
+}
+
+//' Fill selected pixels by interpolation from surrounding areas
+//'
+//' `fillNodata()` is a wrapper for `GDALFillNodata()` in the GDAL Algorithms
+//' API. This algorithm will interpolate values for all designated nodata 
+//' pixels (pixels having an intrinsic nodata value, or marked by zero-valued
+//' pixels in the optional raster specified in `mask_file`). For each nodata 
+//' pixel, a four direction conic search is done to find values to interpolate
+//' from (using inverse distance weighting).
+//' Once all values are interpolated, zero or more smoothing iterations
+//' (3x3 average filters on interpolated pixels) are applied to smooth out 
+//' artifacts.
+//'
+//' @note
+//' The input raster will be modified in place. It should not be open in a
+//' `GDALRaster` object while processing with `fillNodata()`.
+//'
+//' @param filename Filename of input raster in which to fill nodata pixels.
+//' @param band Integer band number to modify in place.
+//' @param mask_file Optional filename of raster to use as a validity mask
+//' (band 1 is used, zero marks nodata pixels, non-zero marks valid pixels).
+//' @param max_dist Maximum distance (in pixels) that the algorithm 
+//' will search out for values to interpolate (100 pixels by default).
+//' @param smooth_iterations The number of 3x3 average filter smoothing
+//' iterations to run after the interpolation to dampen artifacts
+//' (0 by default).
+//' @returns Logical indicating success (invisible \code{TRUE}).
+//' An error is raised if the operation fails.
+//' @examples
+//' ## fill nodata edge pixels in the elevation raster
+//' elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
+//' 
+//' ## get count of nodata
+//' df = combine(elev_file)
+//' head(df)
+//' df[is.na(df$storml_elev),]
+//' 
+//' ## make a copy that will be modified
+//' mod_file <- paste0(tempdir(), "/", "storml_elev_fill.tif")
+//' file.copy(elev_file,  mod_file)
+//' 
+//' fillNodata(mod_file, band=1)
+//' 
+//' df_mod = combine(mod_file)
+//' head(df_mod)
+//' df_mod[is.na(df_mod$storml_elev_fill),]
+// [[Rcpp::export(invisible = true)]]
+bool fillNodata(std::string filename, int band, std::string mask_file = "",
+		double max_dist = 100, int smooth_iterations = 0) {
+
+	GDALDatasetH hDS = NULL;
+	GDALRasterBandH hBand = NULL;
+	GDALDatasetH hMaskDS = NULL;
+	GDALRasterBandH hMaskBand = NULL;
+	CPLErr err;
+	
+	hDS = GDALOpenShared(filename.c_str(), GA_Update);
+	if(hDS == NULL)
+		Rcpp::stop("Open raster failed.");
+	hBand = GDALGetRasterBand(hDS, band);
+	
+	if (mask_file != "") {
+		hMaskDS = GDALOpenShared(mask_file.c_str(), GA_ReadOnly);
+		if(hMaskDS == NULL)
+			Rcpp::stop("Open raster failed.");
+		hMaskBand = GDALGetRasterBand(hMaskDS, 1);
+	}
+	
+	err = GDALFillNodata(hBand, hMaskBand, max_dist, 0, smooth_iterations,
+			NULL, GDALTermProgressR, NULL);
+	
+	if (err != CE_None)
+		Rcpp::stop("fillNodata() failed.");
+		
+	GDALClose(hDS);
+	if (hMaskDS != NULL)
+		GDALClose(hMaskDS);
+		
+	return true;
 }
 
 //' Raster reprojection